[fix] Main Pill 수정
