+++ web_frontend/src/components/SeniorMain.tsx
@@ -1,66 +1,34 @@
 import styled from "styled-components";
-import {
-  AiOutlineCarryOut,
-  AiOutlineMedicineBox,
-  AiOutlineRadarChart,
-  AiOutlineCalendar,
-  AiTwotoneVideoCamera,
-  AiOutlineUser,
-} from "react-icons/ai";
-import { useEffect, useState } from "react";
+import Pill from "./seniorSchedule/Pill";
 
-function SeniorMain() {
-  const [isRecoding, setRecoding] = useState(true);
-  useEffect(() => {
-    setRecoding(true);
-  }, []);
+function SeniorMai() {
   return (
-    <StSeniorMain>
-      <Header>
-        <User>
-          {isRecoding ? (
-            <AiTwotoneVideoCamera color="lightgreen" size="4rem" />
-          ) : (
-            <AiTwotoneVideoCamera color="red" size="4rem" />
-          )}
-          <AiOutlineUser size="4rem" />
-        </User>
-      </Header>
-      <MenuList>
-        <MenuItem>
-          <AiOutlineMedicineBox size="3rem" />
-          복용중인 약<Arrow>{">"}</Arrow>
-        </MenuItem>
-        <MenuItem>
-          <AiOutlineCarryOut size="3rem" />
-          일정 관리<Arrow>{">"}</Arrow>
-        </MenuItem>
-        <MenuItem>
-          <AiOutlineCalendar size="3rem" />
-          캘린더<Arrow>{">"}</Arrow>
-        </MenuItem>
-        <MenuItem>
-          <AiOutlineRadarChart size="3rem" />
-          운동 및 식사 기록<Arrow>{">"}</Arrow>
-        </MenuItem>
-      </MenuList>
-    </StSeniorMain>
+    <>
+      <StHeader>
+        <StUser>아이콘</StUser>
+      </StHeader>
+      <StMain>
+        <StMainItem>
+          <Pill />
+        </StMainItem>
+        <StTodo>할일</StTodo>
+        <StAIDoc>AI</StAIDoc>
+        <StFood>음식</StFood>
+        <StExercise>운동</StExercise>
+      </StMain>
+    </>
   );
 }
 
-const StSeniorMain = styled.div`
-  padding: 0rem 2rem;
-  margin: 0;
-`;
-
-const Header = styled.header`
+const StHeader = styled.header`
+  font-size: 2rem;
   display: flex;
   width: 100%;
   padding-top: 3rem;
-  padding-bottom: 15rem;
+  padding-bottom: 3rem;
 `;
 
-const User = styled.div`
+const StUser = styled.div`
   margin-left: auto;
   font-size: 1.5rem;
   display: flex;
@@ -69,27 +37,28 @@ const User = styled.div`
   }
 `;
 
-const MenuList = styled.ul``;
-
-const MenuItem = styled.li`
-  background-color: white;
-  width: 100%;
-  color: black;
-  display: inline-flex;
-  margin-left: 1rem;
-  margin-bottom: 3rem;
-  margin-top: 3rem;
-  border-radius: 0.5rem;
+const StMain = styled.div`
   font-size: 2rem;
-  &:hover {
-    color: gray;
-  }
 `;
 
-const Arrow = styled.div`
-  width: 3rem;
-  height: 3rem;
-  margin-left: auto;
+const StMainItem = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  padding: 2.4rem 1.6rem;
+  gap: 4rem;
+  flex: none;
+  order: 1;
+  align-self: stretch;
+  flex-grow: 0;
 `;
 
-export default SeniorMain;
+const StTodo = styled.div``;
+
+const StAIDoc = styled.div``;
+
+const StFood = styled.div``;
+
+const StExercise = styled.div``;
+
+export default SeniorMai;
