+++ web_frontend/src/core/router.tsx
@@ -2,7 +2,6 @@ import { BrowserRouter, Route, Routes } from "react-router-dom";
 import SeniorMypage from "../pages/SeniorMypage";
 import SeniorSchedule from "../components/seniorSchedule/seniorSchedule";
 import SeniorMain from "../components/SeniorMain";
-import SeniorMai from "../components/SeniorMai";
 function Router() {
   return (
     <BrowserRouter>
@@ -11,7 +10,6 @@ function Router() {
         <Route path="/schedule" element={<SeniorSchedule />} />
         <Route path="/senior/main" element={<SeniorMain />} />
       </Routes>
-      <SeniorMai />
     </BrowserRouter>
   );
 }
